Remove razzle babel config
