--- .babelrc
@@ -1,28 +1,0 @@
-{
-  "presets": [
-    [
-      "razzle",
-      {
-        "preset-react": {
-          "runtime": "automatic",
-          "importSource": "@emotion/react"
-        }
-      }
-    ]
-  ],
-  "plugins": [
-    ["@emotion", { "autoLabel": "always" }],
-    "@babel/plugin-proposal-optional-chaining",
-    "@loadable/babel-plugin"
-  ],
-  "env": {
-    "development": {
-      "plugins": [
-        ["@emotion", { "autoLabel": "always" }],
-        "@babel/plugin-proposal-optional-chaining",
-        "@loadable/babel-plugin",
-        "react-refresh/babel"
-      ]
-    }
-  }
-}
