+++ package.json
@@ -37,11 +37,14 @@
     "npm": ">=5.3.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.20.2",
-    "@emotion/babel-plugin": "11.10.2",
-    "@emotion/eslint-plugin": "11.10.0",
-    "@emotion/jest": "11.10.0",
-    "@loadable/babel-plugin": "^5.13.2",
+    "@babel/core": "^7.22.5",
+    "@babel/preset-env": "^7.22.5",
+    "@babel/preset-react": "^7.22.5",
+    "@babel/preset-typescript": "^7.22.5",
+    "@emotion/babel-plugin": "11.11.0",
+    "@emotion/eslint-plugin": "11.11.0",
+    "@emotion/jest": "11.11.0",
+    "@loadable/babel-plugin": "^5.15.3",
     "@ndla/scripts": "^1.0.4",
     "@ndla/types-backend": "^0.2.12",
     "@ndla/types-embed": "^2.0.3",
@@ -63,9 +66,8 @@
     "@types/webpack-env": "^1.16.0",
     "@types/webpack-hot-middleware": "^2.25.6",
     "@types/webpack-node-externals": "^2.5.3",
-    "babel-jest": "^29.3.1",
-    "babel-loader": "^9.1.0",
-    "babel-preset-razzle": "^4.2.18",
+    "babel-jest": "^29.5.0",
+    "babel-loader": "^9.1.2",
     "chalk": "^4.1.2",
     "concurrently": "^8.0.1",
     "copy-webpack-plugin": "^11.0.0",
